+++ core/src/main/java/com/ray3k/particleparkpro/Utils.java
@@ -7,6 +7,7 @@ import com.badlogic.gdx.Net.HttpMethods;
 import com.badlogic.gdx.Net.HttpRequest;
 import com.badlogic.gdx.backends.lwjgl3.Lwjgl3Graphics;
 import com.badlogic.gdx.files.FileHandle;
+import com.badlogic.gdx.graphics.Cursor.SystemCursor;
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.ParticleEffect;
 import com.badlogic.gdx.graphics.g2d.ParticleEmitter;
@@ -18,6 +19,7 @@ import com.badlogic.gdx.math.MathUtils;
 import com.badlogic.gdx.net.HttpRequestBuilder;
 import com.badlogic.gdx.utils.*;
 import com.ray3k.particleparkpro.widgets.poptables.PopError;
+import com.ray3k.particleparkpro.widgets.poptables.PopImageError;
 import com.ray3k.particleparkpro.widgets.tables.ClassicTable;
 import com.ray3k.particleparkpro.widgets.tables.WizardTable;
 import org.lwjgl.glfw.GLFW;
@@ -31,6 +33,8 @@ import java.text.DecimalFormat;
 
 import static com.ray3k.particleparkpro.Core.*;
 import static com.ray3k.particleparkpro.Settings.*;
+import static com.ray3k.particleparkpro.widgets.panels.EffectEmittersPanel.effectEmittersPanel;
+import static com.ray3k.particleparkpro.widgets.panels.EmitterPropertiesPanel.emitterPropertiesPanel;
 
 /**
  * A convenience class with various static methods that perform various utility tasks throughout Particle Park Pro.
@@ -100,7 +104,7 @@ public class Utils {
         ShaderProgram.pedantic = false;
     }
 
-    public static void loadParticle(FileHandle fileHandle) {
+    public static boolean loadParticle(FileHandle fileHandle) {
         var newParticleEffect = new ParticleEffect();
         try {
             if (fileHandle.type() != FileType.Internal) newParticleEffect.load(fileHandle, fileHandle.parent());
@@ -112,11 +116,14 @@ public class Utils {
             }
             newParticleEffect.setPosition(0, 0);
         } catch (Exception e) {
-            var pop = new PopError("Error loading particle file. Ensure that all associated images are saved locally.", e.getMessage());
-            pop.show(stage);
+            Gdx.input.setInputProcessor(foregroundStage);
+            Gdx.graphics.setSystemCursor(SystemCursor.Arrow);
+
+            var pop = new PopImageError("Error loading particle file. Ensure that all associated images are saved locally.", e.getMessage(), fileHandle, false);
+            pop.show(foregroundStage);
 
             Gdx.app.error(Core.class.getName(), "Error loading particle file.", e);
-            return;
+            return false;
         }
 
         disposeParticleEffect();
@@ -135,6 +142,7 @@ public class Utils {
                 if (i < emitter.getSprites().size) sprites.put(path, emitter.getSprites().get(i));
             }
         }
+        return true;
     }
 
     public static void disposeParticleEffect() {
@@ -147,24 +155,38 @@ public class Utils {
         }
     }
 
-    public static void mergeParticle(FileHandle fileHandle) {
+    public static boolean mergeParticle(FileHandle fileHandle) {
+        var newParticleEffect = new ParticleEffect();
         var oldEmitters = new Array<ParticleEmitter>();
         var oldActiveEmitters = new ObjectMap<ParticleEmitter, Boolean>();
-        for (var emitter : particleEffect.getEmitters()) {
-            var oldEmitter = new ParticleEmitter(emitter);
-            oldEmitters.add(oldEmitter);
-            oldActiveEmitters.put(oldEmitter, activeEmitters.get(emitter));
-        }
+        try {
+            for (var emitter : particleEffect.getEmitters()) {
+                var oldEmitter = new ParticleEmitter(emitter);
+                oldEmitters.add(oldEmitter);
+                oldActiveEmitters.put(oldEmitter, activeEmitters.get(emitter));
+            }
 
-        if (fileHandle.type() != FileType.Internal) particleEffect.load(fileHandle, fileHandle.parent());
-        else {
-            var textureAtlas = new TextureAtlas(Gdx.files.internal("default/default.atlas"));
-            particleEffect.load(fileHandle, textureAtlas);
-            var defaultImageHandle = Gdx.files.internal("particle.png");
-            fileHandles.put(defaultImageHandle.name(), defaultImageHandle);
+            if (fileHandle.type() != FileType.Internal) newParticleEffect.load(fileHandle, fileHandle.parent());
+            else {
+                var textureAtlas = new TextureAtlas(Gdx.files.internal("default/default.atlas"));
+                newParticleEffect.load(fileHandle, textureAtlas);
+                var defaultImageHandle = Gdx.files.internal("particle.png");
+                fileHandles.put(defaultImageHandle.name(), defaultImageHandle);
+            }
+        } catch (Exception e) {
+            Gdx.input.setInputProcessor(foregroundStage);
+            Gdx.graphics.setSystemCursor(SystemCursor.Arrow);
+
+            var pop = new PopImageError("Error merging particle file. Ensure that all associated images are saved locally.", e.getMessage(), fileHandle, true);
+            pop.show(foregroundStage);
+
+            Gdx.app.error(Core.class.getName(), "Error merging particle file.", e);
+            return false;
         }
 
         activeEmitters.clear();
+        particleEffect = newParticleEffect;
+
         for (var emitter : particleEffect.getEmitters()) {
             emitter.setPosition(oldEmitters.first().getX(), oldEmitters.first().getY());
             activeEmitters.put(emitter, true);
@@ -181,6 +203,12 @@ public class Utils {
         for (var emitter : oldEmitters) {
             activeEmitters.put(emitter, oldActiveEmitters.get(emitter));
         }
+
+        effectEmittersPanel.populateEmitters();
+        effectEmittersPanel.updateDisableableWidgets();
+        emitterPropertiesPanel.populateScrollTable(null);
+
+        return true;
     }
 
     public static ParticleEmitter createNewEmitter() {